feat: devide role of each elements
