--- src/components/darkMode.ts
@@ -1,3 +1,0 @@
-import { writable } from 'svelte/store';
-
-export const darkMode = writable<boolean>(false);
