+++ src/routes/(static)/news/+page.svelte
@@ -3,7 +3,7 @@
     import '@material/web/list/list-item.js';
     import '@material/web/divider/divider.js';
     import '@material/web/list/list.js';
-    import { darkMode } from "../../../components/darkMode";
+    import { darkMode } from "../../../utils/theme/darkMode";
 
     let { data } = $props();
